Work on App console. Segregated MongoSettings, now implemplements in Settings. Created EventWiring interface to bundle together mongo collection and relevant marshallers and unmarshallers. Made 'amount' in events a double.

+++ src/main/java/io/github/tjheslin1/eventsourcedbanking/cqrs/query/DepositFundsUnmarshaller.java
@@ -13,7 +13,7 @@ public class DepositFundsUnmarshaller implements BalanceEventJsonUnmarshaller {
     @Override
     public DepositFundsBalanceEvent unmarshallBalanceEvent(Document document) {
         int accountId = Integer.parseInt(document.get("accountId").toString());
-        int amount = Integer.parseInt(document.get("amount").toString());
+        double amount = Double.parseDouble(document.get("amount").toString());
 
         LocalDateTime timeOfEvent = LocalDateTime.parse(document.get("timeOfEvent").toString(), eventDatePattern());
